feat: refactor validation behavior to use ErrorOr

+++ src/Application/Features/TodoLists/DeleteTodoList.cs
@@ -1,11 +1,10 @@
-﻿using MediatR;
+﻿using ErrorOr;
+
+using MediatR;
 
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
 using VerticalSliceArchitecture.Application.Common;
-using VerticalSliceArchitecture.Application.Common.Exceptions;
-using VerticalSliceArchitecture.Application.Domain.Todos;
 using VerticalSliceArchitecture.Application.Infrastructure.Persistence;
 
 namespace VerticalSliceArchitecture.Application.Features.TodoLists;
@@ -15,26 +14,34 @@ public class DeleteTodoListController : ApiControllerBase
     [HttpDelete("/api/todo-lists/{id}")]
     public async Task<ActionResult> Delete(int id)
     {
-        await Mediator.Send(new DeleteTodoListCommand(id));
+        var result = await Mediator.Send(new DeleteTodoListCommand(id));
 
-        return NoContent();
+        return result.Match(
+            _ => NoContent(),
+            Problem);
     }
 }
 
-public record DeleteTodoListCommand(int Id) : IRequest;
+public record DeleteTodoListCommand(int Id) : IRequest<ErrorOr<Success>>;
 
-internal sealed class DeleteTodoListCommandHandler(ApplicationDbContext context) : IRequestHandler<DeleteTodoListCommand>
+internal sealed class DeleteTodoListCommandHandler(ApplicationDbContext context) : IRequestHandler<DeleteTodoListCommand, ErrorOr<Success>>
 {
     private readonly ApplicationDbContext _context = context;
 
-    public async Task Handle(DeleteTodoListCommand request, CancellationToken cancellationToken)
+    public async Task<ErrorOr<Success>> Handle(DeleteTodoListCommand request, CancellationToken cancellationToken)
     {
         var todoList = await _context.TodoLists
-            .Where(l => l.Id == request.Id)
-            .SingleOrDefaultAsync(cancellationToken) ?? throw new NotFoundException(nameof(TodoList), request.Id);
+            .FindAsync([request.Id], cancellationToken);
+
+        if (todoList is null)
+        {
+            return Error.NotFound(description: "TodoList not found.");
+        }
 
         _context.TodoLists.Remove(todoList);
 
         await _context.SaveChangesAsync(cancellationToken);
+
+        return Result.Success;
     }
 }
